pyproject.toml
[tool.poetry]
name = "file-todo"
version = "0.1.0"
description = ""
authors = ["Valerolactone <varanchukv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
aioboto3 = "^13.1.1"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"
black = "^24.4.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"
pylint = "^3.2.5"
isort = "^5.13.2"

[tool.pytest.ini_options]
pythonpath = [
    '.'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
